refactor: seperating tab button HTML elments from index.js
